delete unused multi spawners
