+++ server/src/main/kotlin/net/horizonsend/ion/server/features/starship/ai/spawning/tsaii/TsaiiSpawner.kt
@@ -1,4 +1,4 @@
-package net.horizonsend.ion.server.features.starship.ai.spawning.pirate
+package net.horizonsend.ion.server.features.starship.ai.spawning.tsaii
 
 import net.horizonsend.ion.server.IonServer
 import net.horizonsend.ion.server.configuration.AISpawningConfiguration
@@ -7,16 +7,15 @@ import net.horizonsend.ion.server.features.starship.ai.spawning.template.BasicSp
 import net.kyori.adventure.text.Component
 import org.bukkit.Location
 
-class PirateSingleSpawner : BasicSpawner(
-	"PIRATE_SINGLE",
-	IonServer.aiSpawningConfiguration.spawners::pirateSingle,
+class TsaiiSpawner : BasicSpawner(
+	"TSAII_ATTACK",
+	IonServer.aiSpawningConfiguration.spawners::tsaiiSingle,
 ) {
 	override fun findSpawnLocation(): Location? = findPrivateerSpawnLocation(configuration)
 
 	override val spawnMessage: Component? = null
 
 	companion object {
-		val defaultConfiguration = AISpawningConfiguration.AISpawnerConfiguration(
-		)
+		val defaultConfiguration = AISpawningConfiguration.AISpawnerConfiguration() // TODO
 	}
 }
